// frontend/src/BookSeries/Details/BookSeriesDetails.css.d.ts
interface CSSModule {
  readonly container: string;
  readonly header: string;
  readonly title: string;
  readonly details: string;
  readonly detailRow: string;
  readonly label: string;
  readonly value: string;
  readonly description: string;
  readonly monitoredIcon: string;
}

declare const styles: CSSModule;

export default styles;